Changed parameters getting in CustomUserView view, added currency convertion to CustomUserView

+++ FundsService/user/models.py
@@ -5,7 +5,7 @@ class CustomUser(models.Model):
 
     user_id = models.AutoField(primary_key=True)
     username = models.CharField(max_length=30, unique=True)
-    funds = models.IntegerField(default=0)
+    funds = models.FloatField(default=0)
 
     def __str__(self) -> str:
         return f'({self.user_id})  {self.username} - {self.funds} UAN'
